@@ -1888,22 +1888,22 @@ GeometryCollection
         >>> s = geopandas.GeoSeries(
         ...     [
         ...         Polygon([(0, 0), (1, 1), (0, 1), (0, 0)]),
-        ...         Polygon([(0, 0), (10, 10), (0, 10), (0, 0)]),
+        ...         Polygon([(0, 0), (0.5, -1), (1, 0), (1, 1), (-0.5, 0.5)]),
         ...     ]
         ... )
         >>> s
-        0       POLYGON ((0 0, 1 1, 0 1, 0 0))
-        1    POLYGON ((0 0, 10 10, 0 10, 0 0))
+        0                      POLYGON ((0 0, 1 1, 0 1, 0 0))
+        1    POLYGON ((0 0, 0.5 -1, 1 0, 1 1, -0.5 0.5, 0 0))
         dtype: geometry
 
         >>> s.maximum_inscribed_circle()
-        0    LINESTRING (0.29297 0.70703, 0.5 0.5)
-        1        LINESTRING (2.92969 7.07031, 5 5)
+        0    LINESTRING (0.29289 0.70711, 0.5 0.5)
+        1    LINESTRING (0.4668 0.25977, 1 0.25977)
         dtype: geometry
 
         >>> s.maximum_inscribed_circle(tolerance=2)
-        0    LINESTRING (0.25 0.5, 0.375 0.375)
-        1          LINESTRING (2.5 7.5, 2.5 10)
+        0    LINESTRING (0.29289 0.70711, 0.5 0.5)
+        1             LINESTRING (0.375 0.25, 0 0)
         dtype: geometry
 
         See Also
@@ -2996,7 +2996,7 @@ GeometryCollection
            :align: center
 
         >>> polygon = Polygon([(0, 0), (2, 2), (0, 2)])
-        >>> s.geom_equals(polygon)
+        >>> s.geom_equals(polygon, align=True)
         0     True
         1    False
         2    False
@@ -3572,7 +3572,7 @@ GeometryCollection
            :align: center
 
         >>> polygon = Polygon([(0, 0), (1, 0), (1, 1), (0, 1)])
-        >>> s.overlaps(polygon)
+        >>> s.overlaps(polygon, align=True)
         0     True
         1     True
         2    False
@@ -3800,7 +3800,7 @@ GeometryCollection
            :align: center
 
         >>> polygon = Polygon([(0, 0), (2, 2), (0, 2)])
-        >>> s.within(polygon)
+        >>> s.within(polygon, align=True)
         0     True
         1     True
         2    False
@@ -4402,7 +4402,7 @@ GeometryCollection
     # Binary operations that return a GeoSeries
     #
 
-    def difference(self, other, align=None):
+    def difference(self, other, align=None, grid_size=None):
         """Return a ``GeoSeries`` of the points in each aligned geometry that
         are not in `other`.
 
@@ -4422,6 +4422,11 @@ GeometryCollection
         align : bool | None (default None)
             If True, automatically aligns GeoSeries based on their indices.
             If False, the order of elements is preserved. None defaults to True.
+        grid_size : float | None (default None)
+            The cell size of the precision grid to use. All the vertices of the
+            output geometry will fall on the grid defined by the grid size.
+            If None, the highest precision (smallest grid size) of the inputs
+            is used.
 
         Returns
         -------
@@ -4511,9 +4516,9 @@ GeometryCollection
         GeoSeries.union
         GeoSeries.intersection
         """
-        return _binary_geo("difference", self, other, align)
+        return _binary_geo("difference", self, other, align, grid_size=grid_size)
 
-    def symmetric_difference(self, other, align=None):
+    def symmetric_difference(self, other, align=None, grid_size=None):
         """Return a ``GeoSeries`` of the symmetric difference of points in
         each aligned geometry with `other`.
 
@@ -4537,6 +4542,11 @@ GeometryCollection
         align : bool | None (default None)
             If True, automatically aligns GeoSeries based on their indices.
             If False, the order of elements is preserved. None defaults to True.
+        grid_size : float | None (default None)
+            The cell size of the precision grid to use. All the vertices of the
+            output geometry will fall on the grid defined by the grid size.
+            If None, the highest precision (smallest grid size) of the inputs
+            is used.
 
         Returns
         -------
@@ -4626,9 +4636,11 @@ GeometryCollection
         GeoSeries.union
         GeoSeries.intersection
         """
-        return _binary_geo("symmetric_difference", self, other, align)
+        return _binary_geo(
+            "symmetric_difference", self, other, align, grid_size=grid_size
+        )
 
-    def union(self, other, align=None):
+    def union(self, other, align=None, grid_size=None):
         """Return a ``GeoSeries`` of the union of points in each aligned geometry with
         `other`.
 
@@ -4649,6 +4661,11 @@ GeometryCollection
         align : bool | None (default None)
             If True, automatically aligns GeoSeries based on their indices.
             If False, the order of elements is preserved. None defaults to True.
+        grid_size : float | None (default None)
+            The cell size of the precision grid to use. All the vertices of the
+            output geometry will fall on the grid defined by the grid size.
+            If None, the highest precision (smallest grid size) of the inputs
+            is used.
 
         Returns
         -------
@@ -4740,9 +4757,9 @@ GeometryCollection
         GeoSeries.difference
         GeoSeries.intersection
         """
-        return _binary_geo("union", self, other, align)
+        return _binary_geo("union", self, other, align, grid_size=grid_size)
 
-    def intersection(self, other, align=None):
+    def intersection(self, other, align=None, grid_size=None):
         """Return a ``GeoSeries`` of the intersection of points in each
         aligned geometry with `other`.
 
@@ -4763,6 +4780,11 @@ GeometryCollection
         align : bool | None (default None)
             If True, automatically aligns GeoSeries based on their indices.
             If False, the order of elements is preserved. None defaults to True.
+        grid_size : float | None (default None)
+            The cell size of the precision grid to use. All the vertices of the
+            output geometry will fall on the grid defined by the grid size.
+            If None, the highest precision (smallest grid size) of the inputs
+            is used.
 
         Returns
         -------
@@ -4853,7 +4875,7 @@ GeometryCollection
         GeoSeries.symmetric_difference
         GeoSeries.union
         """
-        return _binary_geo("intersection", self, other, align)
+        return _binary_geo("intersection", self, other, align, grid_size=grid_size)
 
     def clip_by_rect(self, xmin, ymin, xmax, ymax):
         """Return a ``GeoSeries`` of the portions of geometry within the given
@@ -5367,7 +5389,7 @@ GeometryCollection
     def buffer(
         self,
         distance,
-        resolution=16,
+        quad_segs=None,
         cap_style="round",
         join_style="round",
         mitre_limit=5.0,
@@ -5391,7 +5413,7 @@ GeometryCollection
         distance : float, np.array, pd.Series
             The radius of the buffer in the Minkowski sum (or difference). If np.array
             or pd.Series are used then it must have same length as the GeoSeries.
-        resolution : int (optional, default 16)
+        quad_segs : int (optional, default 16)
             The resolution of the buffer around each vertex. Specifies the number of
             linear segments in a quarter circle in the approximation of circular arcs.
         cap_style : {'round', 'square', 'flat'}, default 'round'
@@ -5442,7 +5464,7 @@ GeometryCollection
             "buffer",
             self,
             distance=distance,
-            resolution=resolution,
+            quad_segs=quad_segs,
             cap_style=cap_style,
             join_style=join_style,
             mitre_limit=mitre_limit,
@@ -6293,9 +6315,6 @@ GeometryCollection
           3  3.0 -1.0
         """
         if include_m:
-            if not compat.SHAPELY_GE_21:
-                raise ImportError("Shapely >= 2.1 is required for include_m=True.")
-
             # can be merged with the one below once min requirement is shapely 2.1
             coords, outer_idx = shapely.get_coordinates(
                 self.geometry.values._data,
@@ -6357,7 +6376,7 @@ GeometryCollection
 
         return pd.Series(distances, index=self.index, name="hilbert_distance")
 
-    def sample_points(self, size, method="uniform", seed=None, rng=None, **kwargs):
+    def sample_points(self, size, method="uniform", rng=None, **kwargs):
         """
         Sample points from each geometry.
 
@@ -6416,14 +6435,6 @@ GeometryCollection
         """  # noqa: E501
         from .geoseries import GeoSeries
         from .tools._random import uniform
-
-        if seed is not None:
-            warn(
-                "The 'seed' keyword is deprecated. Use 'rng' instead.",
-                FutureWarning,
-                stacklevel=2,
-            )
-            rng = seed
 
         if method == "uniform":
             if pd.api.types.is_list_like(size):
